Refactor DRAMBLYS, VORAS, and VILKAS agents into modular packages
Refactored three large agent files into organized package structures to
improve maintainability and readability. All main agent files are now
under 500 lines.

DRAMBLYS (1716 → 17 lines):
- agent.py: Core business logic (1006 lines)
- cli.py: Command-line interface (361 lines)
- validation.py: Word validation utilities (66 lines)
- staging.py: Import staging operations (409 lines)
- display.py: Console output formatting (166 lines)

VORAS (1678 → 17 lines):
- agent.py: Core business logic (774 lines)
- cli.py: Command-line interface (350 lines)
- batch.py: Batch processing operations (226 lines)
- coverage.py: Translation coverage reporting (347 lines)
- display.py: Console output formatting (100 lines)

VILKAS (834 → 17 lines):
- agent.py: Core business logic (688 lines)
- cli.py: Command-line interface (123 lines)
- display.py: Console output formatting (64 lines)

Benefits:
- Separation of concerns: CLI, business logic, and display are isolated
- Easier to test individual components
- More maintainable codebase with focused modules
- Backward compatible: main files are now simple wrappers
