+++ src/wordfreq/agents/dramblys/__init__.py
@@ -1,0 +1,10 @@
+"""
+Dramblys - Missing Words Detection Agent
+
+This package contains the Dramblys agent for detecting missing words
+that should be in the dictionary.
+"""
+
+from wordfreq.agents.dramblys.agent import DramblysAgent
+
+__all__ = ['DramblysAgent']
