+++ src/wordfreq/agents/vilkas/__init__.py
@@ -1,0 +1,10 @@
+"""
+Vilkas - Lithuanian Word Forms Checker Agent
+
+This package contains the Vilkas agent for checking Lithuanian word forms
+in the database.
+"""
+
+from wordfreq.agents.vilkas.agent import VilkasAgent
+
+__all__ = ['VilkasAgent']
