+++ src/wordfreq/agents/voras/__init__.py
@@ -1,0 +1,10 @@
+"""
+Voras - Multi-lingual Translation Validator and Populator
+
+This package contains the Voras agent for validating and populating
+multi-lingual translations.
+"""
+
+from wordfreq.agents.voras.agent import VorasAgent
+
+__all__ = ['VorasAgent']
